+++ src/main/plugins/load-plugins.js
@@ -1,0 +1,7 @@
+import { loadPlugin } from "./load-plugin.js";
+
+function loadPlugins(plugins = []) {
+  return Promise.all(plugins.map((plugin) => loadPlugin(plugin)));
+}
+
+export default loadPlugins;
